Map shared memory pools lazily
Some clients (e.g. Xwayland) send us lots of shared pools and create
buffers from them, but never use the buffers for anything. Since mapping
a pool requires getting some memory from the host and mapping that in,
and freeing the pools requires making a syscall, this can be a
performance problem.

Instead, just hold on to the file descriptor until a buffer is actually
attached to a surface. If it never is, we can just close it without
having to unmap anything.