Don't unscale cursor windows
Low-resolution cursors aren't a big problem and it's better to match the
size of normal Wayland cursors. Also, Vim hides the pointer when you
start typing by setting a 1x1 cursor. When unscaled, this is an illegal
size and Sway hides the cursor, sending Xwayland a leave event. Xwayland
then doesn't bother updating the cursor when Vim tries to show it again.