Initial virtio-gpu support for dmabuf memory
If crosvm is compiled with working minigbm support then allocations are
in video memory and cannot be used with Wl_shm. However, if minigbm
support is not compiled in (or doesn't support the host's hardware) then
crosvm falls back to allocating host memory. Host memory can ONLY be
used with Wl_shm. Therefore, try to detect which case it is and use the
appropriate API. Currently only test.ml uses this, not the proxy itself.

I couldn't get this to work reliably. The scrolling pattern has lots of
glitches, as if some pixels aren't making it to the video memory (needs
a cache flush?). I tried adding some sync calls (`DMA_BUF_SYNC_END`, etc),
but that didn't help, and crosvm keeps crashing with:

    [ERROR:src/linux.rs:2460] vcpu hit unknown error: Bad address (os error 14)

(line number might be a bit out as I added some debug prints)