Simplify code with Host module
Before, the relay connected to the host itself, but accepted a client
connection, which was inconsistent, and `Relay.t` was really just a host
connection. Now `Relay.t` is `Host.t`, which is clearer.

Also, use the new stop functions instead of just shutting down the
transport and waiting for Wayland to notice.